+++ src/gr/aueb/cf/cf9/ch18/bankapp/model/Account.java
@@ -1,16 +1,17 @@
 package gr.aueb.cf.cf9.ch18.bankapp.model;
 
+import java.math.BigDecimal;
 import java.util.Objects;
 
 public class Account {
     private String iban;
-    private double balance;
+    private BigDecimal balance;
 
     public Account() {
 
     }
 
-    public Account(String iban, double balance) {
+    public Account(String iban, BigDecimal balance) {
         this.iban = iban;
         this.balance = balance;
     }
@@ -19,9 +20,13 @@ public class Account {
 
     public void setIban(String iban) { this.iban = iban; }
 
-    public double getBalance() { return balance; }
+    public BigDecimal getBalance() {
+        return balance;
+    }
 
-    public void setBalance(double balance) { this.balance = balance; }
+    public void setBalance(BigDecimal balance) {
+        this.balance = balance;
+    }
 
     @Override
     public String toString() {
